# schema/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FAR-Lab API",
    "description": "FAR-Lab 对外 HTTP API（24§5）。V2 receipts 端点采用统一信封 { ok: true, data: T } + RFC 7807 错误响应（R-05）。",
    "version": "2026-06-27"
  },
...
  "paths": {
...
    "/api/v2/receipts/verify": {
      "post": {
        "summary": "Verify a ProofEnvelopeV2 (six-dimension result)",
        "tags": [
          "v2-receipts"
        ],
        "description": "Verify a submitted ProofEnvelopeV2. Returns unified envelope { ok: true, data: { verification, display } }. Validation errors return RFC 7807 { error_code, message, source_anchor, detail? }.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "schemaVersion": {
                    "type": "string",
                    "minLength": 1
                  },
                  "proofHash": {
                    "type": "string",
                    "minLength": 1
                  },
                  "envelopeId": {
                    "type": "string"
                  },
                  "createdAt": {
                    "type": "string"
                  },
                  "fecHash": {
                    "type": "string"
                  },
                  "claim": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "fecSnapshot": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "protocolFreeze": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "datasetBindings": {
                    "type": "array",
                    "items": {}
                  },
                  "workflowBindings": {
                    "type": "array",
                    "items": {}
                  },
                  "experimentRuns": {
                    "type": "array",
                    "items": {}
                  },
                  "measurementResults": {
                    "type": "array",
                    "items": {}
                  },
                  "statisticalResults": {
                    "type": "array",
                    "items": {}
                  },
                  "verdictTrace": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "antiTheaterReport": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "ledgerRoot": {
                    "type": "string"
                  },
                  "signatures": {
                    "type": "array",
                    "items": {}
                  }
                },
                "required": [
                  "schemaVersion",
                  "proofHash"
                ],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "verification": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "display": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "verification",
                        "display"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error_code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "source_anchor": {
                      "type": "object",
                      "properties": {
                        "fileId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "stageId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        },
                        "callRecordId": {
                          "type": [
                            "null",
                            "string"
                          ]
                        }
                      },
                      "required": [
                        "fileId",
                        "stageId",
                        "callRecordId"
                      ],
                      "additionalProperties": false
                    },
                    "detail": {}
                  },
                  "required": [
                    "error_code",
                    "message",
                    "source_anchor"
                  ],
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}